+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- N/A
+
+## [0.4.0 Alpha] 8-21-2024
+
 ### Added
 
 - External terminal
@@ -17,13 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Settings for changing editor tab size and line height
 - Editor code folding, bracket matching, auto indent, indentation markers
 - Detect indentation size on files
-  - Indent markers
 - Copy/paste files in file tree
+- Autocomplete in the editor
+- Notification system
 
 ### Changed
 
 - Tab context menu options disabled for new files with no path
 - Custom styled scrollbars
+- Improved syntax highlighting look
 
 ### Fixed
 
@@ -38,7 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - If the app crashes, the error outputs to the log file
   - You can open the log file folder from the editor toolbar menu under "Help"
 - Syntax Highlighting
-  - (LIST LANGAUGES HERE)
 - Handling of opening a binary file or unsupported file type. Now shows prompt in editor view telling you so
 - Ability to focus on buttons and inputs via Tab
   - Keyboard support for the file tree is not yet included